RunsMonitor logic was added to RunningContext
